draft
i'm not exactly familiar with collaborative github yet so i dunno what i'm doing but i've put my thoughts on how the subject classes/objects should maybe look like in draft.js.

Also, idk if i should be committing or creating pull requests. Pull requests might be smarter so we could talk through the changes, but I'll be committing this one since it's simply a rearrangement of files and a new draft file.